requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.9
aiohttp-retry==2.9.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
apispec==6.8.2
APScheduler==3.11.0
async-timeout==5.0.1
attrs==25.3.0
backports-datetime-fromisoformat==2.0.3
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
exceptiongroup==1.3.0
Flask==3.1.1
Flask-HTTPAuth==4.8.0
flask-marshmallow==1.3.0
Flask-Pydantic==0.13.1
flask-swagger-ui==5.21.0
frozenlist==1.6.2
google-auth==2.40.2
google-genai==1.17.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
marshmallow==4.0.0
mistune==3.1.3
multidict==6.4.4
numpy==2.2.6
opencv-python==4.11.0.86
outcome==1.3.0.post0
packaging==25.0
pillow==11.2.1
propcache==0.3.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.10.1
PySocks==1.7.1
pytesseract==0.3.13
python-dotenv==1.1.0
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.25.1
rsa==4.9.1
selenium==4.33.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.30.0
trio-websocket==0.12.2
twilio==9.6.2
typing-inspection==0.4.1
typing_extensions==4.13.2
tzlocal==5.3.1
urllib3==2.4.0
webargs==8.7.0
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.20.0
